+++ event-store-management/event-store-management-core/src/main/java/uk/gov/justice/services/eventstore/management/replay/process/EventBufferProcessorFactory.java
@@ -3,7 +3,7 @@ package uk.gov.justice.services.eventstore.management.replay.process;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessorProducer;
 import uk.gov.justice.services.event.buffer.api.EventBufferService;
-import uk.gov.justice.services.event.sourcing.subscription.error.StreamErrorService;
+import uk.gov.justice.services.event.sourcing.subscription.error.StreamErrorRepository;
 import uk.gov.justice.services.event.sourcing.subscription.manager.EventBufferProcessor;
 import uk.gov.justice.services.event.sourcing.subscription.manager.cdi.InterceptorContextProvider;
 
@@ -18,7 +18,7 @@ public class EventBufferProcessorFactory {
     private InterceptorChainProcessorProducer interceptorChainProcessorProducer;
 
     @Inject
-    private StreamErrorService streamErrorService;
+    private StreamErrorRepository streamErrorRepository;
 
     @Inject
     private InterceptorContextProvider interceptorContextProvider;
@@ -29,7 +29,7 @@ public class EventBufferProcessorFactory {
         return new EventBufferProcessor(
                 interceptorChainProcessor,
                 eventBufferService,
-                streamErrorService,
+                streamErrorRepository,
                 interceptorContextProvider,
                 componentName);
     }